feat: api refactor, docker image push action
