+++ frontend/src/api/bucket.ts
@@ -1,0 +1,15 @@
+import { mapKeysToCamelCase } from '@app/utils';
+import { API_URLS } from '.';
+
+async function getBucketDetails(slug: string) {
+  const response = await fetch(`${API_URLS.library.buckets}/${slug}`);
+
+  if (!response.ok) {
+    throw new Error('Failed to fetch bucket details');
+  }
+  const data = await response.json();
+
+  return mapKeysToCamelCase(data);
+}
+
+export { getBucketDetails };
