+++ frontend/src/app/(app)/library/[ladder]/[bucket]/page.tsx
@@ -1,29 +1,7 @@
 import { Breadcrumbs } from '@app/components/modules/Breadcrumbs';
-import { mapKeysToCamelCase } from '@app/utils';
 import { BucketDetails } from '@app/components/modules/BucketDetails';
-import { API_URLS } from '@app/api';
-
-async function getBucketDetails(slug: string) {
-  const response = await fetch(`${API_URLS.library.buckets}/${slug}`);
-
-  if (!response.ok) {
-    throw new Error('Failed to fetch bucket details');
-  }
-  const data = await response.json();
-
-  return mapKeysToCamelCase(data);
-}
-
-async function getLadderName(slug: string) {
-  const response = await fetch(`${API_URLS.library.ladders}/${slug}`);
-
-  if (!response.ok) {
-    throw new Error('Failed to fetch ladder details');
-  }
-  const data = await response.json();
-
-  return mapKeysToCamelCase(data).ladderName;
-}
+import { getLadderName } from '@app/api/ladder';
+import { getBucketDetails } from '@app/api/bucket';
 
 export default async function BucketDetailed({ params }: { params: { bucket: string; ladder: string } }) {
   const { bucket, ladder } = params;
@@ -43,3 +21,5 @@ export default async function BucketDetailed({ params }: { params: { bucket: str
     </div>
   );
 }
+
+export const dynamic = 'force-dynamic';
