+++ frontend/src/static/icons/ChevronDoubleLeft.tsx
@@ -3,9 +3,9 @@ export const ChevronDoubleLeft: React.FC<React.SVGProps<SVGSVGElement>> = (props
     <path
       d="M10 16L6.53034 12.5303C6.23745 12.2374 6.23745 11.7626 6.53034 11.4697L10 8M17 16L13.5303 12.5303C13.2374 12.2374 13.2374 11.7626 13.5303 11.4697L17 8"
       stroke="#A1A5AC"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
+      strokeWidth="1.5"
+      strokeLinecap="round"
+      strokeLinejoin="round"
     />
   </svg>
 );
