+++ frontend/src/static/icons/CloseIcon.tsx
@@ -1,5 +1,5 @@
 export const CloseIcon: React.FC<React.SVGProps<SVGSVGElement>> = (props) => (
   <svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
-    <path d="M0.75 0.75L15.25 15.25M15.25 0.75L0.75 15.25" stroke="#555C6B" stroke-width="1.5" stroke-linecap="round" />
+    <path d="M0.75 0.75L15.25 15.25M15.25 0.75L0.75 15.25" stroke="#555C6B" strokeWidth="1.5" strokeLinecap="round" />
   </svg>
 );
